Delete `GlobalResolution.is_instantiated`
Summary:
I was going to try to figure out exactly what this function
does and come up with a better name, and then I realized it's
dead code anyway. Delete it.

I still need to figure out `ClassHierarchy.is_instantiated`
but that's less urgent since it's a pure function rather than
part of the environment stack (currently I view environment
stack functions - which tend to have a lot of redirection that
makes it slower to find and read the source code - to be a
higher priority in terms of having really clear names.)

Reviewed By: grievejia

Differential Revision: D52301739

fbshipit-source-id: 457ad92ddc3bd011762d26403784c94693cc6591