Rename and comment on class_hierarchy_contains_class
Summary:
This function was named is_tracked, which is a pretty ambiguous name
because we also talk about "tracked" modules a lot.

What it really does is just call `class_exists` in a very indirect way
that might just auto-return `true` if `EnvironmentControls.t` says so.

We use it to (partially) support retaining the names of untracked types
in no-daemon-query, which is a hack to help with glean indexing. My
hope is that we can eventually replace this with some notion of untracked
types that preserve their names (if Pyre is intended as an IDE tool, and not
just a type checker, then it is not good behavior to just drop the names of
types we can't find), but in the meantime I at least want to (a) have the most
unambiguous possible name for this function and (b) have a comment explaining
how exactly it differs from `class_exists`, and why that matters.

Reviewed By: grievejia

Differential Revision: D52297258

fbshipit-source-id: 0ea6e07e33a778b13f6abe22af217a9bd3005be9