Standardize GlobalResolution signatures [1]
Summary:
For predictability, we really want all the GlobalResolution functions to have
consistent call signatures. Right now most of them take the resolution value as
a first positional argument, but some of them used named arguments and we
aren't even consistent about the argument name (sometimes it is `resolution`,
sometimes `global_resolution`).

This is both pointlessly unpredictable and a bit of a drag on refactoring,
so I'm going to make it all positional.

There are a lot of functions to change so I'll stack it into a few
chunks; this diff handles all the ones that were using a named
`~global_resolution` argument.

Reviewed By: grievejia

Differential Revision: D52307908

fbshipit-source-id: d2582071c660939870db6e301950883ea5ccfe26