Move the type-contains-untracked-name logic
Summary:
This logic is only used inside of aliasEnvironment. It should live there
rather than directly in unannotatedGlobalEnvironment.

In addition, the name of the function could be clearer so I renamed it,
and the use of `annotation` for both a type and all of the identifiers inside
the type was offending my sensibilities so I fixed that.

Reviewed By: connernilsen

Differential Revision: D52297646

fbshipit-source-id: ce5d26d77b3c649503aded5720b851c9bea19834