Standardize one last GR function signature
Summary:
After my previous sweep through GlobalResolution standardizing
signatures, almost all functions take a `GlobalResolution.t` as
the first argument, which makes skimming and refactoring easier.

The `annotation_parser` is the one exception; because this function
returns a "thing" used to parse expressions later, it has only an
optional argument aside from GlobalResolution, which leads to
"can't erase optional argument" issues if we try to put the positional
argument first.

This is one reason why as a general rule of thumb it's good to put named /
optional arguments first, but in the special case of an environment / analysis
API function I want to prioritize standardized interfaces over that rule of
thumb, because a standard interface that allows currying (which the optional
argument prevents) can make refactors easier.

To that end, I get rid of the optional arg in the standard way: make
two functions to handle the validating vs nonvalidating cases.

Reviewed By: kinto0

Differential Revision: D52433465

fbshipit-source-id: d06c38e46cc6933f41016859575485296a016276