module_path_of_qualifier
Summary:
Before making a pass at extracting the ModuleTracker and AstEnvironment
apis into a new abstract API (which Jia and I discussed as the first
step of enabling a pyre-parse buck action), I want to make sure the
names of the functions are all as descriptive as possible.

So I'm going to make a quick stack renaming the functions - whose names
are currently kind of vague and redundant - to instead reflect the
kinds of things taken as inputs vs outputs.

This is the first rename: `lookup_module_path` becomes
`module_path_of_qualifier`. I chose this over `qualifier_to_module_path` to
follow typical OCaml naming conventions for conversion functions.

Reviewed By: kinto0

Differential Revision: D52516281

fbshipit-source-id: 69661c17838e86188cf072fb4f1c5f07a897580a