Add basic support for is_generated
Summary: Right now ig manually filters this out by manually filtering on file names, they have requested that generated code not be included in roll ups and other tools in the dashboard. This is the first step in getting that data into the pipelines and then into the tool itself. We should also be able to index on it, which should make filtering this data faster and simpler from both a code and user perspective.

Reviewed By: inseokhwang

Differential Revision: D52545219

fbshipit-source-id: cecd68dd48cd79107db81648bf6256dcd0322f2d