Extract a Parsing module from AstEnvironment
Summary:
This commit extracts the "IncomingDataComputation"
portion of the AstEnvironment API into a "Parsing" module,
and also moves the `ParserError.t` type definition there.

This Parsing module is pure output data types + logic, there's
no reference to a concrete environment. It takes the controls
and the get_raw_source function as input.

Reviewed By: grievejia

Differential Revision: D52516279

fbshipit-source-id: 3e84f541827f01a69a3cb231596c33aba2de2977