Rename a couple more low-level functions
Summary:
Continuing my theme of making it easier to skim names of apis
to see how they fit together, renaming some loading-related
functions to describe the input and output value, e.g.
`get_raw_code` -> `code_of_module_path`.

This started partially because I was tempted to try to make `get_raw_code` take
a qualifier as input which is what I'd sort of expect based on the name, but it
turns out that's not a great idea (the `Source.t` actually needs some metadata
found in the `ModulePath.t` anyway).

Reviewed By: grievejia

Differential Revision: D52543011

fbshipit-source-id: 848bcede258a044fbf2851f68d96046a07c83c9e