project_qualifiers -> type_check_qualifiers
Summary:
The notion of "project qualifiers" is pretty vague, it is supposed to indicate
"module names of the user's source code but not dependencies."

But (a) the only thing we actually use this for is deciding what to type check,
and (b) Pysa type checks dependencies too, which makes the name misleading in
that context.

Naming it for what we actually want makes everything a little more obvious.

Reviewed By: grievejia

Differential Revision: D52543756

fbshipit-source-id: 79306b44f6561acfeaff148e876a5c1ee25167a2