Simplify Parsing and clean up datatypes
Summary:
In diff review of D52604867, Jia suggested that Parsing
could just expose pure functions without callbacks for code loading
if we just pipe the load result through.

I agree, that's a really nice change. But it's not quite trivial because one of
the points of this code is converting load errors into parse errors, so that
the rest of Pyre doesn't have to worry about the distinction (which is not a
useful one in anything downstream).

This commit makes the change, which isn't quite trivial, by:
- Defining loadResult.t and ParseResult.t datatypes in Parsing, which
  really makes all the type signatures much clearer
- Changing our outermost "get the source" function to take a load result
  as input, instead of injecting the load function as a callback
- Renaming that outermost function to `parse_result_of_load_result` in
  order to make the name as clear and consistent as possible.

In the process I'm able to get rid of the "functional dependency" part of
the module level docstring that prompted the suggestion to simplify.

Reviewed By: grievejia

Differential Revision: D52609440

fbshipit-source-id: 5fc6d4b185c4700cd1e227e1866359f90bd08e02