Fix nondeterministic test by using `dict.fromkeys ~= orderedset` (#814)
Summary:
Pull Request resolved: https://github.com/facebook/pyre-check/pull/814

**Problem**

For a number of weeks now I've been frustrated by github test failure
notifications; many of my commits are generating tasks to fix test failures.

This test is actually disabled in internal CI because it isn't deterministic; that means at least it isn't noisy internally, but we should still fix it!

**Root Cause**

I dug into the root cause for BE day, and it is that we
are using `list(set(_))` to deduplicate two lists of references in
the find-references logic.

Deduping is totally correct, but also nondeterministic because `set` ordering
isn't fixed.

**Solution**

There's no ordered set in the python standard library, and
it's not worth adding a dependency to get one. Instead, the standard trick
to work around this is to use a dict of `None` so the keys act like a set,
which solves the problem because in 3.7+ dict keys are ordered.

Reviewed By: inseokhwang

Differential Revision: D52662624

fbshipit-source-id: 8e8b94c89dda2f3f5768b9607ed1f2854570f8a0