Use descriptive names for PathLookup functions
Summary:
The `module_of_source_path` name was okay, but we generally use
"qualifier" when talking about module names as `Reference.t` values.

The `instantiate_path` name was more problematic; the notion of
"instantiation" already appears in a lot of places like type variables,
so we probably want to avoid overloading the term when there's a good
alternative.

The solution here is to rename everything using `{result_type}_of_{input_type}`
style names as the ocaml standard library usually does.

Reviewed By: inseokhwang

Differential Revision: D52671552

fbshipit-source-id: eea487d0c5a98ac710fe1999a582838ec0bcb26b