Namespace the ModuleTracker code to handle ArtifactPaths
Summary:
To improve the modularity of the interfaces we expose, it's helpful
to somehow demarcate parts of the existing API that only make sense
in certain contexts.

One such chunk of the API is the part of ModuleTracker.ReadOnly that
deals directly with ArtifactPaths; this isn't really a part of the core
API used by the rest of the analyzer.

Instead, it is *only* useful in contexts like a top-level pyre server
where we have a BuildSystem and we need to use ArtifactPath.t as the
intermediate data when converting from
```
qualifier (module name) -> ArtifactPath.t -> SourcePath.t
                        |                  |
									ModuleTracker        BuildSystem
```
and vice-versa.

So, I just made a new namespace `ModuleTracker.ReadOnly.ArtifactPaths`
to make this context clearer.

Reviewed By: kinto0

Differential Revision: D52673737

fbshipit-source-id: dfdc88682d49b4912cddd26fb94cb814faef2f05