Fix a bad copy-pasted name in `SourcePath.ml`
Summary:
The SourcePath.t and ArtifactPath.t types currently
are exactly the same representation (they are different
in *meaning*, but the code is the same).

We actually created them by defining it and copy-pasting
the implementation.

This is mostly not a problem, but I forgot to rename
the data constructor which is very confusing!!

Fixing that :)

Also fixing license formatting; a codemod to add warning
suppressions is making the linter angry.

Reviewed By: inseokhwang

Differential Revision: D52672026

fbshipit-source-id: f86ded5b258273e497115d930e519b417a26a982