Improve control flow in classic overlay updates
Summary:
The logical control flow for an overlay update is:

- get a SourcePath.t, convert it to an ArtifactPath.t in the build system
- convert that to a ModulePath.t in ModuleTracker
- process everything downstream in terms of the ModulePath's qualifier

That's exactly what we do in the environment stack today, but currently
the request handler is throwing away the module information and then
re-computing it using ModuleTracker.

This creates a diamond-like control flow, which is harder to follow than
if we just extract the qualifiers from the update result. So in this diff
I make that change, which makes the control flow linear :)

Reviewed By: inseokhwang

Differential Revision: D52671048

fbshipit-source-id: 7303309ff2f6e34be3ed2556e7951b82f43fefd3