Remove `all_module_paths` from ModuleTracker.ReadOnly
Summary:
This function is not at all a part of the logical ModuleTracker API; it is used
exclusively for global all-or-nothing (as opposed to incremental)
invalidation of cached type environments for Pysa.

The Pysa caching is not something we want to break since it's important for
developer productivity (it's not used in production, only local work), but
leaking this into the ReadOnly API confuses things.

Fortunately Pysa is already using a raw `ModuleTracker.t` here, so the change
is entirely internal, the interface already banned access through read-only

Reviewed By: grievejia

Differential Revision: D52806602

fbshipit-source-id: 303d65de12cdeeaf5506efdcca6fb6a51895549f