Better logging for unresolved callees
Summary:
Currently we do not log why a callee is unresolved, which makes it difficult to
debug. This diff proposes to log the reason of creating `unresolved`
`CallCallees` when flag `debug` is true.

An alternative is to log such reason regardless of whether `debug` is true, but
this would create too much log.

Another alternative is to store the reason inside `CallCallees` and then output
to `callgraph.json`, which is also not ideal because (1) this uses more memory,
since the reasons are strings, (2) it would be more difficult to see the
reasons, because it requires a user to download the result bundle.

Reviewed By: arthaud

Differential Revision: D52711624

fbshipit-source-id: 9d624618086eda2c05f1d39854e1353ffe77420b