Reproduce unresolved calls to @abstractclassmethod
Summary:
Context: We are no longer resolving calls to methods decorated with
abstractclassmethod, due to a recent typeshed upgrade, which are hence now
introduced to function `typeshed_stubs` inside `test.ml` (where the typeshed
code is quite out of date...).

This diff provides a minimal repro of the above problem.

Reviewed By: arthaud

Differential Revision: D52697003

fbshipit-source-id: 9fea97db2e1317fe2d03346745e0b848fd4b1b67