Kill dead global-listing hooks in ScratchProject
Summary:
I'm working on a refactor to split core APIs that are part of analysis -
all of which should be O(request) operations such as getting type info
needed in a single type-check-define operation - from "global" operations
that can list all of the modules in a project and use that to get "all"
of some kind of thing.

One of the places I may have to update all the apis is in `test.ml`,
but before I do that I want to eliminate dead code paths that are global,
we have quite a few functions defined that are never used!

Reviewed By: grievejia

Differential Revision: D52823411

fbshipit-source-id: 868d6b226adb9d14f5d30811342f65aaada4d2a3