Use GlobalModulePathsApi in test.ml
Summary:
The necessary scratch project interfaces (type_check_qualifiers and
project_sources) are re-implemented in terms of the global module paths
api.

Moreover, populating a type environment is split into a separate helper
function that returns a unit because it makes the code in
`build_type_environment` significantly easier to follow.

Reviewed By: kinto0

Differential Revision: D52847216

fbshipit-source-id: b064e601e1bc3993becc3ad2aa54a6415747218b