Use GlobalModulePathsApi in UnannotatedGlobalEnvironment (start)
Summary:
This diff begins the process of stripping global functions from
UnannotatedGlobalEnvironment.ReadOnly by creating a new sub-module
for global queries, where the queries require a
GlobalModulePathsApi as input.

The old apis remain present because there are quite a few callsites to fix
and I want to update a few at a time in small, easy-to-review diffs; once
callsites are all updated we'll strip out the old functions.

This diff only updates the code itself and the two easiest callsites:
- query.ml, where we already have a global module paths api in scope
- the unit tests, where we can get one from our scratch project

Reviewed By: kinto0

Differential Revision: D52847217

fbshipit-source-id: 3b7c3a96e531c0c252c93df01c203b6b1bc58e97