Resolve callees additionally for parametric class types when ignoring decorators
Summary:
# Problem
With the recent typeshed upgrade, Pysa can no longer resolves calls to callees
that are decorated with abstractclassmethod.

# Solution
This diff enables resolving the calls to methods that are decorated with
abstractclassmethod, by resolving the callee type while pretending the
decorators do not exist. In particular, this diff adds a new pattern matching
case to match against calls where the types of the base expressions (at the
call sites) are class types that are parametric (such as the test added in the
previous diff).

Reviewed By: alexkassil

Differential Revision: D52875238

fbshipit-source-id: 8bf02a80fce9c31699c073b17658c92825af6885