Limit third-party vendoring, part 1: the script
Summary:
We currently vendor the entire `typeshed/stubs` directory of typeshed,
which is not desirable for a number of reasons:

- It increases the size of our vendored typeshed, but there's an accepted
  way to get third party stubs: use the `-stubs` packages, which internally
  could be buckified and included in TARGETS. This has speed implications.
- Unlike the `-stubs` packages, you can't version different interfaces based
  on the library version so it's actually broken in the face of backward-
	incompatble api changes on a major version upgrade of a library.
- It leads to an enormous `search_paths` in our backend because each library
  becomes its own search root. This is a (small but nonzero) performance hit
	in module tracking, and more importantly it introduces a ton of line noise
	in our logging; I'm constantly frustrated by the line noise when debugging
	unrelated issues

Getting rid of third-party vendoring entirely would be involved and really
isn't all that important; having it for a handful of core libraries is annoying
but not a big problem if the libraries are very stable (you can kind of think
of them as an "extended stdlib").

So, to improve matters greatly without all the work of a full migration, for
now I'm using an allowlist to limit our vendoring to just whatever we know is
needed. This greatly reduces the existing problem, and also prevents things
from getting worse as the upstream `typeshed/stubs` library continues to
grow moving forward.

This diff is only the script change because if I combine the script change with
the actual typeshed changes it exceeds phabricator UI limits and you won't be
able to see the script in diff review. So, I'll stack a "part 2" on top that
actually regenerates the typeshed using the allowlist.

Reviewed By: grievejia

Differential Revision: D52916513

fbshipit-source-id: deb9571eadac217fc0cb3d63297292d9b5717729