Remove dead TODO, rename private function
Summary:
We added propagation of root updates to overlays long ago in
`run_update_root`, so the TODO is dead.

We still expose the ability to update only the root and inspect
the result in unit tests, but by renaming that in the public API
we can make it safer.

Reviewed By: kinto0

Differential Revision: D52964698

fbshipit-source-id: 223a89db56d57ddf88bf5c5c58787aeb63cf4660