Move ArtifactPaths to stand-alone module
Summary:
Based on review suggestions not to put ArtifactPath-base apis
directly into the (new in an upcoming diff) SourceCodeApi
module, I put them in a submodule of ModuleTracker.ReadOnly.

But these apis are necessary for the current code, and as a result
we need to be able to call them after hiding ModuleTracker from
other layers of the code, which is a short-term goal.

The solution is to pull them out of ModuleTracker.ReadOnly and into
a separate ocaml module; this will make it easy to flip the underlying
`module_path_of_qualifier` logic from `ModuleTracker.ReadOnly` to
`SourceCodeApi`, but without polluting the api itself.

That strikes a balance between the need to migrate *existing* code away
from direct use of `ModuleTracker` versus the desire to not force
ArtifactPath logic into place in a buck-controlled world where
ModulePath.relative_path may be all we need.

Reviewed By: grievejia

Differential Revision: D52969798

fbshipit-source-id: fa64c7455af711bcfe045fab72d4141d5dfbf321