Eliminate read-only AstEnvironment/ModuleTracker from environment functor
Summary:
To unblock an initial buck-invokes-pyre implementation for per-target type checking,
we want to fully abstract away the details of ModuleTracker and AstEnvironment. This
is needed because it is much simpler to create an alternative, share-nothing implementation
of the ModuleTracker api than to adapt the existing implementation.

The rafactor to do this will be straightforward but very large. A good step before we begin
actually flipping types is to eliminate interfaces we don't want to keep from the functor.
It's nice to do this first because the functor automatically threads these types through
the entire codebase; only by making this change can we easily start cutting over apis
one at a time.

This commit is a pure refactor of the functor interface in `environment.ml`, no
actual public interfaces change anywhere in the stack of realized layers; we
can recover the interfaces stripped from the functor by extracting them through
`unannotated_global_environment`.

In this diff we leave the `ModuleTracker.Overlay` handles in place, I'll get rid of those
in a separate diff to keep changes as focused as possible.

Reviewed By: grievejia

Differential Revision: D52969797

fbshipit-source-id: 16b346989f73356675a60471a7c22c857ce89879