Enable workspace/symbol requests
Summary:
Update fbcode/tools/pyre/client/commands/daemon_querier.py to call the helper method and hook it up to fbcode/tools/pyre/client/commands/pyre_language_server.py

Task can be found here T174171520 and this code is for item #7 in the task. Previous diffs in this stack are for items 1 to 6.

Reviewed By: inseokhwang

Differential Revision: D52896722

fbshipit-source-id: 56d50d63696fb600e69296d3a154d45333947446