Emit field data as feature for Legacy GraphQL XIDs
Summary:
# Problem
We can't tell during XID resolution on legacy graphql responses what the field is
# Solution
Add a feature that specifies the field and object in the pattern `via-graphql-legacy-field:GraphQLObject.ObjectField`
e.g. `via-graphql-legacy-field:GraphUser.pronouns`

In order to access this data, we must first move the logic in `gather_functions_to_model` to `compute_models` so we can use the info we need on the graphql object.

Explored making the FormattableAnnotation more generic and injecting it into `__str__` in `generate_taint_models/model.py` but decided it's not worth it for the first use I can find of a dynamic breadcrumb

Reviewed By: arthaud

Differential Revision: D52881459

fbshipit-source-id: 266c05f9626f959832fc9a0912ffd3a363480e45