introduce SourceCodeApi
Summary:
Create a SourceCodeApi module, representing the combination of a
(read-only) ModuleTracker and AstEnvironment.

At the moment there are no use cases, but you can create one of these using
either an `AstEnvironment.ReadOnly.t` or an
`UnannotatedGlobalEnvironment.ReadOnly.t`.

The handles to do this are *explicitly* tracked or untracked (with that
explicit in the name of the function), which I hope will help rectify a
type safety gap in our current code, where it is too easy to forget the
`?dependency` in read-only calls and accidentally break module tracking.

The eventual goal of this stack will be to *completely* abstract
away both `AstEnvironment` and `ModuleTracker`, which will require not
only the existing `SourceCodeApi` core read-only logic, but also an
"abstracted" dependency-tracked updatable API with `Base` and `Overlay`
interfaces.

The stack will proceed by first replacing all the uses of read-only
AstEnvironment and ModuleTracker with the existing `SourceCodeApi.t`,
and only then attempting to replace the read-write interfaces with something
abstracted.

Reviewed By: grievejia

Differential Revision: D53009418

fbshipit-source-id: a5ca61d16119d11590a63d1d7035a9792270d78f