Flip read-only errors environment to SCA
Summary:
This stack kicks off an effort to decouple the environment stack
from direct access to AstEnvironment and ModuleTracker, replacing
them with SourceCodeApi.

This diff flips the read-only errors environment, and updates callsites;
most of the changes amount to re-wiring the top-level of daemons
to use the new abstract API.

Because the interface change goes a bit viral due to code sharing (basically
once we flip `ArtifactPaths`, we have to flip all callsites) we also have
to add `source_code_api` creation hooks to `TypeEnvironment.ReadOnly`. But
we defer deleting ast_environment and module_tracker access until the next diff.

Reviewed By: grievejia

Differential Revision: D53009413

fbshipit-source-id: 806c6a914cf6eb3a05694e3097869a21d7c0ce15