Add .mli file for SourceCodeApi
Summary:
I like to defer adding the interface file in a stack until
I've largely cut over callsites, because if the api needs adjusting
it's a pain to have to edit two files every time.

At this point we've cut over pretty much all the read-only use cases
(outside of tests at least) to SourceCodeApi, so I'm confident in
the interface.

Reviewed By: grievejia

Differential Revision: D53009416

fbshipit-source-id: caf969344c96a2a256c2d9a84d838dda75faac8b