Push `global_module_paths_api` up
Summary:
I implemented `global_module_paths_api` as a hook on a read-write
ModuleTracker initially in an effort to just split global queries off
from the core read-only environment logic.

At this point though, because we want to abstract ModuleTracker and
AstEnvironment in one shot, we want to push most such accesses up
so that they run through UnannotatedGlobalEnvironment + AstEnvironment.

This aligns better with the needs of our refactor, where this will become
(for now, until we can refactor to the point that we actually differentiate
lazy vs eager environments) a hook on the abstract SourceCodeIncrementalApi.

Reviewed By: grievejia

Differential Revision: D53090474

fbshipit-source-id: a46ccfe492c01242304da3245157d550f70255fd