Begin working on SourceCodeIncrementalApi
Summary:
In order to quickly unblock per-target type checking (by quickly, I mean
without finishing an end-to-end refactor of the environment stack that
modularizes dependency tracking *everywhere*, which is something we want
but don't need yet and will take time), we decided to move forward with
an approach where we abstract the implementation of the existing
dependency-tracked, incremental AstEnvironment and ModuleTracker as they
are.

This means that in addition to the simple read-only `SourceCodeApi`
with core read-only logic we also need a `SourceCodeIncrementalApi`
module providing abstract versions of the full interface:
- a dependency-aware read-only interface, that returns the "normal"
  read only handle but in the tracked case takes a dependency
	as input
- a base module with the standard filesystem-backed updatable api
  and a handle to read-only
- an overlay module with ownership hooks

This diff starts on the process by introducing the ReadOnly module,
and moving the existing get-a-tracked-or-untracked-api logic from
AstEnvironment.ReadOnly to SourceCodeIncrementalApi.ReadOnly.

The Base and Overlay modules are stubbed out for now, I want to fill
out these interfaces as I cut over use cases because it's a little hard
for me to clearly see exactly what they should be in advance.

Reviewed By: grievejia

Differential Revision: D53090477

fbshipit-source-id: 0a513a3de4df4904cb8ddff2d7422f1a8c1e18e5