Strip module_tracker from environment interfaces
Summary:
This diff finishes stripping ModuleTracker entirely from the interfaces
of all environment layers from UGE and above. As a result, there are also
no references to ModuleTracker whatsoever above UGE, even in the implementations.

To do this, the pysa cache logic that accesses ModuleTracker (which really is using
ModuleTracker, because it's directly relying on the serialization) has to go
by way of UGE -> ast environment -> module tracker to get there.

Reviewed By: grievejia

Differential Revision: D53090476

fbshipit-source-id: 64575f85339d68c56c5568cad9da3c2f35c6a721