Rename test hooks that assume AstEnvironment
Summary:
In order to emphasize that in general one should not assume there is an
underlying AstEnvironment and ModuleTracker, I've nested the test hooks
into `ScratchProject.ReadWrite.AssumeBackedByAstEnvironment`.

Once we've finished enabling an abstract `SourceCodeIncrementalApi`,
this naming will emphasize that in general you cannot assume there is
an ast environment inside but for some use cases (tests of the underlying
modules + pysa tests that rely on the specific implementation) we will
need to continue making that assumption.

Reviewed By: grievejia

Differential Revision: D53090480

fbshipit-source-id: 5c253e2bc77be06d878421af4db789c66e43adec