Move update types to SourceCodeIncrementalApi
Summary:
Move the definition of the code update type out of ModuleTracker and into
SourceCodeIncrmentalApi; ModuleTracker is now just *implementing* an abstract
interface, so the type needs to live in the interface definition.

We haven't yet defined an Overlay.t, but just by moving the type definition
this diff completes the decoupling of several key modules from AstEnvironment
and ModuleTracker:
- `environment.ml` needed the update type but has no other references to
  bottom layers
- the same is true for `overlaidEnvironment.ml`

For similar reasons, moving the UpdateResult type out of AstEnvironment and
into SourceCodeApi now is helpful; both Base and Overlay depend on this type
in the update API, so without moving it ahead of time we can't refactor
Base and Overlay in separate diffs.

Reviewed By: grievejia

Differential Revision: D53090478

fbshipit-source-id: 4aadf8fc05b877dc05b3a2581d752d96f3cedc5c