Rename the UGE handle to get an AstEnvironment.t
Summary:
Note: this diff completes the stripping of access to ModuleTracker and
AstEnvironment from UnannotatedGlobalEnvironment (and therefore all
later layers as well).

What remains is to actually cut over the implementation of UGE to
a SourceCodeIncrementalApi interface.

---------------

As I was working toward a clean refactor where UGE and layers above
rely only on SourceCodeIncrementalApi but not the specific
AstEnvironment-based implementation of it, I realized we'll hit a few
snags:
- Lots of unit tests of AstEnvironment in particular create a scratch
  project and then extract the underlying environment, which is reasonable
	given that they are unit tests specifically of that low-level layer.
- Saved state relies directly on the underlying shared memory implementation
  of AstEnvironment and on the ability to push ModuleTracker state into
  shared memory and then serialize.
- The Pysa cache managing logic wants to directly make use of full
  state dumps and mutation of SharedMemory in ModuleTracker and AstEnvironment,
	so they really do need the raw environment rather than an abstract API.

These problems are definitely solvable with more refactoring, (e.g.
we could force AstEnvironment tests to only use the abstract API, we can
build a new PysaEnvironment that encapsulates everything Pysa needs including
the underlying AstEnvironment so that they aren't trying to extract it
from a TypeEnvironment anymore).

But we don't want the refactoring scope that blocks per-target type checking to
creep; we do want to make all these refactors in the long run but they can't
all get done before we make progress on new functionality.

As a result, UnannotatedGlobalEnvironment is going to need a callback
that returns an AstEnvironment, which will be used in specific settings where the
environment is known (not at the type level) to be "classic", i.e. backed by an
AstEnvironment. In a non-AstEnvironment-based situation, that callback should
just raise an exception.

For this reason, I really want to make the name of the hook to invoke that callback
clear, so that no one is tempted to try using it inside the core type check
logic. This diff makes the rename.*

This is a little dirty, but it's really no dirtier than existing hacks such as
making lazy environments fail if you try to list all modules. Again, we can
and should clean this up but it will take more extensive refactors.

Reviewed By: grievejia

Differential Revision: D53090475

fbshipit-source-id: 11469afeff74e2d507a025eb10aaca30c90c6918