Don't suppress dune test command
Summary:
I'm not sure why we have been suppressing the test command,
it seems actually helpful to know the raw dune command in some
cases.

More importantly, I think it helps with understanding how long things
are taking; I was motivated to tweak this because I ran make and it
felt like it was taking a long time on the
```
dune build install -j auto --profile dev
```
and I started wondering whether eden was in a bad state or I had
runaway processes eating my server capacity, and only after a
minute did I realize that it was actually running tests (which
are frequently much slower than the main build).

Reviewed By: grievejia

Differential Revision: D53110381

fbshipit-source-id: 944009303c075b617ca38eccf307ff98001c847f