Parse breadcrumb string literals
Summary:
Right now, `Via` and `ViaDynamicFeature` except an identifier.
This is a problem when the breadcrumb needs to have a dash, since `a-b` is not an identifier, but the expression `Minus(Identifier a, Identifier b)`.

This diff allows to specify breadcrumbs as string literals, i.e `ViaDynamicFeature["foo"]`. This allows for any character in the breadcrumb name.

Reviewed By: tianhan0

Differential Revision: D53185796

fbshipit-source-id: b705e42e882ee29e5d88dc6c368d7132331736c7