Inject is_qualifier_tracked to SCA
Summary:
I found yet another bug where the interface of how to construct
a SourceCodeApi was too simple; I thought that is_qualifier_tracked
was redundant because you could derive it from `module_path_of_qualifier`,
but it turns out that this is not correct:
- Pyre semantics say only explicit qualifiers have ModulePath values
- But `is_qualifier_tracked` is supposed to include implicit (namespace
  package) qualifiers

So, we need to make this injectable, and use the implementation in
module_tracker. Note that any alternative implementation needs to be
aware of this, but hopefully the fact that there are two distinct
callbacks required by `create` will remind us to read the docs and
realize that implicit qualifiers are where the trickiness pops up.

----------

Note about safety of previous diffs: It turns out there is *exactly*
one place in Pyre where implicit qualifiers matter, and that is resolving
direct imports of modules (you can't actually do anything with such a module,
it just "is there" so you can import it).

As a result, this bug doesn't manifest on the current stack because we are
still using module tracker directly from UGE in the relevant codepath. I found
the bug because (in a stacked diff) I was trying to cut over UGE.

-----------

**Does it have to be this messy?**

No. At this point in the stack I'm focused on correctly modeling existing APIs in SoucreCodeApi, but by changing the types we can simplify.

Based partly on suggestions in review of this diff, I was able to change the types to eliminate the need for two separate functions for explicit vs implicit modules in D53158276, replacing both `is_qualifier_tracked` and `module_path_of_qualifier` with a single function `look_up_qualifier`.

Reviewed By: grievejia

Differential Revision: D53137974

fbshipit-source-id: 2114394061c99db2d25ce69caa3116bca93e4d42