Rename `Unsafe.*` submodules to `AssumeXyz`
Summary:
At one point there was only one kind of unsafety to worry about
in Pyre, which was grabbing a read-write inner environment, which
in general can break dependency tracking but is useful in some
contexts where we either know it's okay (Pysa, where no incremental
updates will happen) or in tests. So I decided to mark that operation
as "Unsafe" by sticking it in a submodule of the read-write environment.

These days there are two kinds of not-type-safe operations, and soon there will
be three:
- (existing) access to inner read-write, which can break incremental
- (existing) operations that assume a nonlazy module tracker
- (upcoming) operations that assume an AstEnvironment backs the current
  environment stack, such as saved state serialization.

So I think it's better if the not-type-safe operations are namespaced by what
assumptions you make when calling them, hence I changed the name to "assume". I
originally had these modules named as `UnsafeAssumeXyz`, but I decdided the
`Unsafe` is just redundant - I think `Assume` already pretty clearly states
that you are assuming something that the type signature doesn't guarantee, and
the "Unsafe" is just extra noise.

This commit:
- renames the two preexisting "unsafe" modules to state assumptions
- creates a new such module for accessing global module paths apis, which
  was previously in the top-level interface but is in fact not type safe
	and belongs in a named assumption-tagged sub-module.

Reviewed By: grievejia

Differential Revision: D53140995

fbshipit-source-id: d56ca359b3356510333e94d616b5413ec70f0f40