Push owns_qualifier to AstEnvironment.Overlay
Summary:
In preparation for making the overaly API abstract, we need to get
rid of the part of UGE where we unpack a `ModuleTracker.overlay` from
an ast environment overlay; that code cannot be made abstract.

The quick fix is to push owns_qualifier to the ast environment overlay
interface.

NOTE: as of this diff, there are no direct uses of the `ModuleTracker`
interface from UGE. That means we are (finally!) in a place where this
refactor can actually be a simple mechanical swap of concrete API to
abstract API.

Reviewed By: grievejia

Differential Revision: D53140992

fbshipit-source-id: 2b9d6c3b3ce4ad55a47a31771c547ebf1224008d